gyms/TauGym/requirements.txt
gymnasium
numpy
pyyaml
pytest

# For tau-bench integration
litellm 